tweak the way files are deleted

+++ app/src/main/kotlin/com/simplemobiletools/gallery/activities/MainActivity.kt
@@ -200,17 +200,17 @@ class MainActivity : SimpleActivity(), DirectoryAdapter.DirOperationsListener {
         }
 
         Thread({
-            if (needsPermissions) {
-                val document = getFileDocument(file.absolutePath, config.treeUri)
+            if (!file.delete()) {
+                if (needsPermissions) {
+                    val document = getFileDocument(file.absolutePath, config.treeUri)
 
-                // double check we have the uri to the proper file path, not some parent folder
-                val uri = URLDecoder.decode(document.uri.toString(), "UTF-8")
-                val filename = URLDecoder.decode(file.absolutePath.getFilenameFromPath(), "UTF-8")
-                if (uri.endsWith(filename)) {
-                    document.delete()
+                    // double check we have the uri to the proper file path, not some parent folder
+                    val uri = URLDecoder.decode(document.uri.toString(), "UTF-8")
+                    val filename = URLDecoder.decode(file.absolutePath.getFilenameFromPath(), "UTF-8")
+                    if (uri.endsWith(filename)) {
+                        document.delete()
+                    }
                 }
-            } else {
-                file.delete()
             }
         }).start()
     }
